installer: split run.rs into run, run_context, spawn, file_actions
