+++ installer/src/uninstall.rs
@@ -8,7 +8,8 @@ use anyhow::Result;
 use crate::action_log::action_log_path;
 use crate::install_record;
 use crate::plan::{self, Action, Plan, Stage};
-use crate::run::{self, Ctx};
+use crate::run;
+use crate::run_context::Ctx;
 use crate::setup::self_install::PATH_MARKER;
 use crate::systemd_service;
 use crate::wizards::unitree::g1::CDDS_MARKER;
